Renamed package exception to exceptions
